CharacterFlags: Remarkable Athlete
The Fighter (Champion) ability Remarkable Athlete allows for
prof
bonus rounded up to be added to initiative and str and dex ability
checks where half proficiency is selected for that skill.

This does not add half proficiency to those skills, just rounds up the
half bonus if it is set.