Allow a spellcaster with a spell level of zero
Consider e.g. an Ultroloth, which is a CR13 creature with innate spells 
but no spellcaster levels. At present, unless the user sets the 
spellcasting level to `1` or `-1`, they will have a list of useless 
spell slots added to their NPC sheet (setting it to `1` reduces this,
but does not eliminate it).