Remove "0/0 uses" from compendium items
Switch to `value=null` and `max=""` as these are the values which result
from manually clearing the fields in the UI.

Using 0/0 instead of something null/falsy confuses the Better Rolls
module, which is really a Better Rolls bug, but, since it can be solved
by making the SRD data nicer, we should prefer to improve the data
instead.

Update uses.value and uses.max defaults in template.json to match their UI defaults.

Commit built items.db.