Update spell-icon-migration.json
Migration entries for `icons/magic/light/explosion...` are incorrect.

They are trying to convert paths that do exist such as `explosion-star-large-orange-purple` into paths that do not exist `explosion-star-orange-purple-large`. This results in broken icon paths whenever the migration runs.

The core icons should probably be renamed so alike colours are next to each other and the schema is consistent. This is the case with other icons such as `beam-rays-orange-small.webp` where the size is at the end. Until that time, this change amends the migration to work with the current icon naming scheme.