Add Coin support in _simplifyDiceTerms
Closes #2281
There was an issue in _simplifyDiceTerms where a Coin was reduced to its operator and number of faces, then assumed to be a die when passed back out. This ended up turning a coin into a die 2 and combining it with any other die 2s in the Dice Terms.
This was fixed by using a c for the faces value if the DiceTerm is Coin and returning a DiceTerm of Coin, instead of Die, if the face value is c.
(Any DiceTerm other than Coin should default to using the number of faces and return as a Die DiceTerm)