Move data-type attribute for Empty Spellbook
Changes in PR #2156 moved data-type and data-level attributes to the spellbook-header li tag from the item-create a tag for spellbook's with sections. However for the spellbook-empty li the attributes were still on the a tag. This was causing the _onItemCreate in base-sheet.mjs to get an empty dataset and undefined type. This was fixed by moving those attributes to the li tag for the empty spell book as well.
Closes #2476