Update character-sheet-2.mjs
There was an issue in character-sheet-2.mjs _renderOuter() when mapping a callback we return a new Tabs5e using the this.options.tabs which used the initial as the active, which overwrote the previous active. New code changes the initial to the previous active tab if the initial does not match the active tab, before passing it to the constructor. Additionally changed a check condition in setting GetData() activeTab where the DOM element length was checked and otherwise "details" was set as the activeTab even when we had an existing active tab. This element check was removed.

Closes #2868