[#2939] Fix duplicate D&D logo being added to sidebar when popped out
Caused because the original implementation used the `#game-details` selector
on the document to fetch the element to modify, but when the sidebar is
popped out that ID is no longer unique.

Modified it to instead use the HTML passed in to the `renderSidebar` hook
while guarantees that only the proper sidebar is modified.