[#2843] Fix concatenation issue for AC bonuses
Adds a hardcoded list of fields that should be treated as formula fields
which aren't part of the actor's data model. Currently that list only has
the AC bonus.

Also had to create a fake FormulaField instance in order to properly
case the AE value into something that can be used in a formula.