[#2867] Add context menus to race, background, classes, & subclasses
Adds a context menu on ActorSheet5eCharacter2 targeting any element with
`data-item-id` that is inside `.pills-lg`, which allows it to target all
the concept items that exist outside the normal inventory elements.

In order to properly activate the context menus on these elements it fetches
the `InventoryElement` on the features tab and uses its `_onOpenContextMenu`.