[#2876] Fix context menus being displayed off the bottom of the screen
Previously the menu sizing wasn't available because it didn't yet
exist in the DOM. Moving the `appendChild` call to the top of the
method resolved that.

This also required anchor the context menu by its bottom edge when
expanding up to ensure the display/hiding animation occurred in the
correct direction.