[#2975] Fix rapid calls to TokenRing#flashColor ending in the wrong color
The `flashColor` method fetches the starting color at the beginning so
it knows where to return to when the animation is complete. The original
implementation fetched this directly from the token ring's data, but
the animation would then modify that data to perform the effect.

In order to fix this, this change now fetches the token ring color from
the token document's flag and `TokenDocument5e#getRingColors` to ensure
that it will always return back to the proper value regardless of its
current color.