[#2993] Fix unidentified items on group sheets showing description
The expand action in `InventoryElement` was calling `Item#getChatData`,
which didn't have an handling of unidentified items. This switches it
over to using `ItemDataModel#getCardData` which does so long as that
method exists. If it doesn't exist (which might happen for item types
provided by modules), it falls back to the old implementation.

Also removes a leftover listener for the old expand function on groups.