[foundryvtt#2952] better centering of icons in header buttons
using font-size-12 or font-size-13 seems to be the root cause of the issue. I think it just can't center an icon of that size nicely in the available space. Using the default font-size-14 resolves the issue.

The width: 1.3em shrinks the padding around the icon to compensate for the larger icon size. I think having the larger icons with less padding is probably fine.