[foundryvtt#2224] auto resize height only on tab change
previously we forced the height of the item sheet to certain values whenever the app was moved (due to overriding setPosition). This approach only automatically changes the height when the tab changes, rather than any time the app moves.

We can also just rely on the standard auto height sizing to adjust the size of the sheet for the content (like for a class or subclass sheet)