Update custom elements with adoptedCallback
For custom elements that use constructed CSS, we have to implement
the adoptedCallback method and manually reconstruct and reattach
a new CSS object to the shadow DOM. This is because when a custom
element is moved to a new document, any constructed CSS objects
that it has adopted already are silently stripped.

We also need to update how we construct our CSS objects. Specifically
we need to ensure we are accessing the current owner document of our
custom element, and not bind the root document from which ever window
we defined our custom element in.

We also can no longer use a static variable to store our CSS obj,
since in Chrome based browsers the static element is shared across
windows (though not the case in firefox). So we need to keep a map
from document to CSS obj singleton and do the lookup manually.