[#1019] Refactor rest methods, add durations, add hook
Reworks the private `Actor5e#_rest` method to take `RestConfiguration`
and `RestResult` as it parameters to allow the full configuration data
to be passed to the completion hooks. A deprecation path has been added
that detects if the old parameters are present and builds these objects
so things don't fail if modules were calling this method.

Rather than looking for separate `longRest` and `shortRest` methods on
the system data, it now just checks for a single `rest` method.

A pair of new hooks, `dnd5e.shortRest` and `dnd5e.longRest`, have been
added that are called just after the configuration dialog has been
completed but before any additional resting action has been taken.