[#2934] Fix issue with active effect de-duplication
First fix was a bug that was thrown if the `origin` on an effect
was `null`. The existing check assumed all origins were strings,
which would throw an error on `null` origins and prevent checks
against subsequent effects from occuring.

The second fix was due to the way the effect de-duplicate worked
with the `persistSourceMigration`. The de-duplication operated by
calling `deleteEmbeddedDocuments` on the document being migrated
to remove the effects, but when `persistSourceMigration` was set
it would perform an update on the actor data without diffing
resulting in the effects being added back. To fix this we swap
to removing the effects from the source data if that flag is set.
This fix is only necessary for compendium migrations because
world migrations perform the update before the de-duplication.