[#3149] Fix advancement with class restriction on NPCs
Because NPCs didn't have `originalClass` in their data model, they
wouldn't get any advancement marked "Original Class Only". Rather
than adding `originalClass` to their data models this changes
`Item5e#isOriginalClass` to return `null` if `originalClass` doesn't
exist. Then the `Advancement#appliesToClass` getter checks if it
is `null` and applies it as if it is the original class. This should
maintain the behavior on PCs that have the proper data and fix it
on NPCs.