[#2920, #3147] Add magical bonus to ammo, fix attack bonus data
Calls `Roll.replaceFormulaData` on the ammo attack bonus before it
is added to the weapon's to hit formula to avoid errors with string
terms not resolving.

Adds `magicalBonus` to ammunition that acts like that on weapons.
This adds to the attack bonus for the weapon when it is being used
with the ammunition. The damage bonus is appended to the first damage
formula on the weapon, and the damage properties gain any physical
properties provided by the ammunition so silvered or magical ammo
properly carries through to the weapon damage.

This also changes how additional ammo damage is added to make use
of the new system for multiple rolls on damage. Ammunition damage
is now constructed as separate rolls with their own type data rather
than being appended to the base weapon damage formula.