[#2668] Derive NPC details from race item if present
Introduces a new `AbilitiesFields#prepareRace` method that takes a
race item and applies its stats to the actor, either complementing
the default values or replacing them depending on whether the `force`
option is set.

This then replaces the existing code in `CharacterData` that used
to apply race data and is added to `NPCData`.