[#893] Add system for selecting summoning placement
The `TokenPlacement` class right now is small and only delegates
to the `TokenPlacementTemplate` which is responsible for rendering
and returning placement information. With the V12 improvements to
placeables hopefully we'll be able to shift more logic out of the
`MeasuredTemplate` subclass and into the `TokenPlacement` class.

The `TokenPlacementConfiguration` data structure currently just
includes the prototype token information, but will eventually have
quantity, origin, and range values to handle multiple summons and
restricting range from the summoner.

Works with tokens of any size or scale on the square grid. On hex
grids, it handles 1x1 tokens pretty good with only a bit of offset
on the final token placement. Stranger token sizes lead to some
issues with positioning, but I'm not sure how important those are
to fix before the other grid improvements in V12.