#3065 Modify Deterministic Removal in simplifyRollFormula
There was an issue where our simplifyRollFormula Deterministic check would loop through the roll terms and remove any term marked as Deterministic. This caused an issue if a *, /, or % operator was used on a Deterministic term, as the Deterministic term was removed but not the operator and other term that can not be evaluated as those operators are dependent on the Deterministic term. The solution was to treat the operator and other term as Deterministic as well, and remove them. This does not affect *, /, and % used on constants.
Additionally modified _enrichRollTooltip in chat-message.mjs to pass _formula as this was the roll's original formula before evaluation where parenthesis are collapsed, providing a more accurate constant.
Closes #3065