[#3183] Adjust rollDamage to return merged roll, rather than first
In `3.0.0` we modified `Item5e#rollDamage` to support separate rolls
for multiple damage parts and added `returnMultiple` option for
backward compatibility. The issue with that initial implementation
is that if `returnMultiple` is `false`, it returns the first roll
made rather than matching previously behavior of having all the
damage parts merged into a single roll.

This change alters the bahavior of `damageRoll` if `returnMultiple`
is `false` and there are more than one roll to create a new merged
roll from the evaulated rolls and return that.

The separate rolls are still stored in the chat message and used by
the system, but the combined rolls are now available to modules that
rely on the `dnd5e.rollDamage` hook or call `Item5e#rollDamage`
directly.