[#3248] Invert healing automatically in `calculateDamage`
If a `DamageDescription` with the type `healing` is encountered in
`Actor5e#calculateDamage` it will now be automatically inverted
rather than relying on the `multiplier` option. This simplifies
the usage of `applyDamage` and opens up the possibility of mixing
damage and healing (if you wanted to do a risky healing feature
that does 1d4 bludgeoning damage and heals 1d6, that is now possible).

This inversion is applied by default, but can be turned off by the
new `invertHealing` option passed to `applyDamage`.

Another new option is `only`, which tells the damage system to only
apply damaging or healing damage parts. This option should be useful
for reworking the damage application interface to allow for three
options: "Apply All", "Apply Damage", & "Apply Healing".