[#3250] Aggregate damage types before passing to `applyDamage`
Adds a new dice helper function `aggregateDamageRolls` which splits
and re-combines damage rolls based on the roll's type, term flavor,
and optionally any roll properties. This function is then used by
`ChatMessage5e#applyChatCardDamage` before rolls are totaled to
ensure that damage types are properly split.

The function starts by using a `chunkTerms` helper function which
goes through the terms in a roll grouping into chunks based on the
operators. Addition and subtraction cause new chunks to be created
while multiplication and division are included in the same chunk.
This is so that a damage type defined in flavor of a term can spread
to other terms that it is multiplied or divided with. So if you have
a formula `1d6 + 5 * 1d4[fire]` that has the overall type of piercing,
the `1d6` will remaining piercing while the `5 * 1d4` will all be
treated as fire.

Once the chunking is complete then chunks from all different rolls
are combined into lists of terms grouped by type and properties
(so long as the `respectProperties` option is set, otherwise they
will only be grouped by type). These sets are then converted into
new `DamageRoll`s with options set appropriately.