[#1372] Add display of resistances, etc. in damage application
Displays what resistances, immunities, vulnerabilities, and
modifications are modifying the damage to be applied. Each of these
can be clicked on to ignore that source of change or, if it is an
immunity, downgrade it. The total damage that will be applied is
updated as you click these buttons just like the multiplier
buttons below.