[#3326] Add option to create spell scrolls with reduced description
Adds a configuration object that is passed to `createScrollFromSpell`
to allow for more customization of the creation process. Right now it
has a single option `explanation` which controls how much of the
spell scroll rules are included in the created scroll.

- `full`: Existing functionality including entire spell scroll rules
- `reference`: Adds scroll level, reference link to rules, and
  potentially a concentration note to beginning of description
- `none`: Just includes the original spell description

In order to support `reference` mod a non-level-specific version
of the spell scroll rules have been added to the magic items section
of the SRD rules and a reference of `&Reference[Spell Scroll]`.