[#3321] Allow spell scrolls to be created at increased level
The `SpellScrollConfiguration` object now has a `level` parameter
that causes the created spell scroll to include a spell at a higher
level than normal. This stores the increase level information in
`flags.dnd5e.spellLevel`, which contains the original leve, the
casting level, and the scaling information.

`Item5e#rollDamage` has been modified to now check for the new flag
before scaling damage, and use that information if present. The
scaling information is stored in the flag because otherwise it
would be lost when converting to a consumable type.