[#1856, #3342] Allow UUIDs to be used as consumption targets
Items with the consumption modes of ammo, item uses, and resource
will now be display their target as a text input field which accepts
an item UUID. When the item is then added to an actor, it will
search for an item on the actor with a matching `sourceId` and set
that item's ID as its consumption target. This should allow for
pre-configuring consumption in compendium items.