[#3345] Add Place Members buttons to group to place into scene
Includes modifications to the `TokenPlacement` class to fully
support multiple token placements. The functionality of that
system has been changed slightly so right clicking will now skip
the current placement, but not cancel the whole process. This
should allow skipping certain group members when placing the group.

There is a current limitation that prevents the already placed
tokens from displaying in the scene.