[#3361, #3362] Add support for NPC hit dice based on HP formula
Calculates the hit dice max and denomination based on a NPC's HP
formula with a fallback based on their size if necessary. On NPCs
`HitPointAdvancement` will now determine the die size based on the
NPC's hit die denomination rather than what is specified in the
class.

Adjust the `Actor#rollHitDie` method to be able to handle rolling
hit dice on NPCs and the resting methods to be able to properly
recover NPC hit dice.