[#3242] Add ability to change creature type when summoning
Introduces a new `creatureTypes` set in summoning configuration
that defines creature types that the summoned creature will be
turned into upon summoning. If more than one type is listed, then
the player will have the chance to choose that in the usage dialog.

Makes use of Foundry's new `<multi-select>` element for input.