[#3374] Add enchantment active effect that modifies an item
Active Effects created with `flags.dnd5e.type` of `enchantment`
will now apply to the item in which they are included, not the
actor.

The existing item sheets will now identify fields that have been
modified by an enchantment effect and disable them, similar to
how field are disabled in actor configs.