[#3176] Powerful Critical Preserve flavor
There was an issue where the Powerful Critical was created by summing all the die faces times number, and adding a numerical term to the terms. It appears this was trying to add Powerful Critical as the type however due to a bug no flavor was getting added. This was modified to use a Map to sum all dice of matching types and add a Numeric term with matching flavor for each. This later allows the aggregate-damage-rolls.mjs's chunkTerms to aggerate the powerful critical rolls to their matching type as non flavored values are added to the default damage roll type.
Closes #3176