[#3374] Add enchantment feature type, improve effects tab
Adds a new feature type "Enchantment" with sub-types matching the
two existing enchantment types in class features: "Artificer
Infusion" and "Rune". Also adds a new "Enchant" action type.
When either of these types are selected a new section appears
in the effects tab for "Enchantments" that allows for creating
a new enchantment active effect.

Two new sections are added for other items listing "Active" and
"Inactive" enchantments to keep them separate from other effects.

Added a new metadata property to `ItemDataModel` for `enchantable`,
which determines whether enchantment active effects will be applied
to this item. Items that are not enchantable can still hold
enchantment active effects but they will not be able to modify the
item.