[#3418] Add enchantment configuration & restrictions
Adds a new `EnchantmentData` embedded data model onto `feat` items
that contains configuration data for enchantments. This includes
a number of helper methods to determine if an item represents an
enchantment or enchantment source.

For enchantments, restrictions are added for whether the enchantment
can be applied to magical items and what types of items it can apply
to. There is also a `canEnchant` method that takes an item and
determines whether it meets the required criteria.

For enchantment sources there is a formula for the maximum number of
enchantments in effect at a time and the period at which enchanted
items can be re-bound.