[#3242] Add level limits to summoning profiles
Adds a minimum and maximum level to each summoning profile hidden
beneath an "Additional Settings" tray. These values determine
whether the profile will be visible for a given spell, character,
or class level.

For summoning on a spell item, it will always use the spell's level.
On all other items it uses the character level by default unless a
class identifier is entered, in which case it will use that class's
level.

The ability use dialog has been updated to dynamically refresh the
list of summoning profiles as the casting level is changed.