[#3305] Enforce enchantment restrictions when dropping onto item
When adding an enchantment to an item that accepts enchantments,
the restrictions specified in the enchantment source will now be
enforced. It will also prevent the creation of an enchantment
effect directly on an actor.

Enchanted items in the world will register themselves with a static
registry of all enchanted items so the enchantment items can easily
determine what items they have enchanted.

Support has been added for items that might offer enchantments or
receive them (to support a weapon that can enchant other weapons
for example). This uses the `origin` data to track whether the
enchantment effect is applied to the item or not. To this end the
default origin for new enchantments isn't set like for normal
effets. The origin is only set when an enchantment is dragged to
another item. This ensures that the enchantment origin always
matches the item doing the enchantment, even if that item was
duplicated.