[#2299] Add replacement option for ItemChoiceAdvancement
Migrates the `configuration.choices` data structure on item choice
advancement to be an object with a new `replacement` boolean. This
indicates that players can replace a previous choice at that level.

The configuration app for `ItemChoiceAdvancement` has been modified
into a three column layout to prevent it from getting too tall when
there are a lot of choices such as for Warlock.

A data model has been added for item choice with a new `replaced`
object that indicates which previous choice was replaced for each
level. The previous choices are kept within the `value.added`
structure to accurately track choices across levels. The flow UI
will display replaced choices as crossed out and has a radio button
next to other previous choices so players can choose which of them
to replace.