[#1787] Support potentially unlinked spells in spell lists
Adds a new `unlinkedSpells` array to spell list journal pages for
holding information on spells that may potentially become unlinked.
This is useful for when a spell list might need to include spells
that aren't in the current module or in the SRD.

In order to properly display, this data structure contains the name
of the spell, its level, and its school. It also contains the
source object for tracking where the spell originated. This source
object can contain a UUID so if the module that contains the spell
is active it will display linked like the rest, but if the module
is inactive then it will display as a static entry.