[#2773] Display limited spell uses on sheet
Adds the uses as the first column on the player's spellbook tab.
These uses only display when the spellbook is wide enough and are
editable like in the features tab.

If the sheet it too narrow, then the uses are instead displayed in
the spell's subtitle alongside and required components. In the
narrow mode they are not editable.