[#3305] Add enchantment usage dialog & application tray
Introduces enchantment to the `AbilityUseDialog` using the
`applyEnchantment` and `enchantmentProfile` options, with similar
details to how summoning is handled.

When enchantment is performed it adds a new "Enchant" tray to
chat cards. This tray has a place to drop an item which is then
enchanted when the apply button is pressed.

Enchantments are tracked with concentration, so if concentration
on the enchanting spell ends then the enchantments will be removed.