[#3512] Allow enchantments to be limited by level
Adds `min` and `max` level limits as flags to the enchantment
effects that are used to limit when those enchantments appear. This
is based off spell level if it is a spell, or either character or
class level for other items depending on whether a class identifier
is provided.

If the "Enchantment Prompt" option is not checked, then the system
will take level restrictions into account when selecting which
enchantment to provide. This should make it easy to configure a
feature like "Enhanced Defense" artificer infusion which swaps
enchantments depending on artifier level, without requiring the
player to view the ability use dialog everytime they want to use
the infusion.