[#3526] Apply dnd5e2 journal styles in journal editor
Updates the journal styles to appear within the journal editor.

Because the new journal styles are based on a class added by the
new `JournalSheet5e` and not by the pages themselves, it was
necessary to add a rendering hook that detected if the journal
the page belongs to uses the new system sheet and then adds the
necessary class to the editor.

Currently the styles are scoped to apply to all journal page
editors, rather than just text pages where they are really needed.
If this is undesired, we can add an additional `.text` class to
the selectors so they only apply when appropriate.