[#3511] Add additional effects that ride along with enchantments
Enchantments can now define one or more effects that will be
added whenever the enchantment is added and removed when the
enchantment is removed. These are configured in the enchantment
config.

In order to prevent any effects used as ride along effects from
being transfered to the actor when not necessary, they get the
`rideAlong` flag which is removed when applied. This prevents
these effects from being transfered to the actor or being offered
in the chat card. This separate flag allows for transfer and
non-transfer ride along effects.