[#3291] Add setting to control chat card tray expansion
Adds a new setting allowing setting the initial expanded state of
damage & effect application trays in chat cards. The settings are
"Expand All" and "Collapse All" plus "Collapse Older", which
expands newly created cards but collapses any cards more than five
minutes old.

To better support the expanding trays, a new CardTrayElement has
been added which adds support for the `open` attribute on card
trays to explicitly control its behavior. This attribute matches
the behavior of the `open` attribute on `<details>` elements, so
adding or removing the attribute will open and close the tray, and
toggling the `open` property will cause the attribute to be added
or removed. The tray also fires off a `toggle` event that matches
the one fired by `<details>`.