[#3544] Fix issue with extra entries in registry when summoning
Resolved an issue being caused by extra entries being added to the
registry when the summoning process was happening due to data
being prepared for a token before the final token with ID was
created.

Also added a `_onDelete` handler to `Token5e` to properly handle
when tokens are deleted, not just actors.