[#2111, #3551] Split attunement & attuned, add optional attunement
Splits `system.attunement` into two values, `attunement` which
takes a string to indicate what kind of attunement is required,
and `attuned` which is a `boolean` indicating whether the item
is currently attuned.

This change allows simplification of a lot of the logic for
attuning/un-attuning and for resetting attunement for dropped
items.

This also adds the "Optional Attunement" mode that is used for the
Spell Gem items and some 3rd party content. If attunement is
optional then active effects and magical bonuses won't be
suppressed if the item isn't attuned.

Also fixes a bug causing `_prepareItems` from being called twice
on the new character sheets.