[#3305] Allow "Enchant Prompt" to still enchant if prompt is shown
Unchecking "Enchant Prompt" is useful when there is only a single
enchantment available at any given time to help players to avoid
seeing the ability use dialog if no other configuration is
required, but if the dialog is shown then it requires the player
to click "Apply Enchantment" or else the enchantment area won't
appear in chat. For a more streamlined approach it would probably
be simpler to just remove that checkbox so the enchantment area
is always shown in chat if there is enchantments available. In
this case the ability use dialog is only necessary when there is
more than one enchantment profile to select.