[#3305] Remove Enchant Prompt option & simplify logic
Seeing as the prompt option for Enchantments wasn't achieving much,
this simply removes it and always shows the ability use dialog if
there is more than one available profile. This should make the
process cleaner and ensure users don't have to fret over the value
to use for that prompt checkbox.

Now the enchantment will always be displayed in the chat message as
long as there is more than one enchantment available (though
a module can use the `dnd5e.preUseItem` hook to prevent that from
happening) and will not be required if there is only a single
enchantment.