[#3558] Fix poison damage icon, allow damage icons to be colored
Adds `--icon-fill` variable to allow for the color on damage icons
to be set, with the default color being black to match the former
coloring. This also creates a duplicate of `poisoned.svg` with
black as the default color rather than white.

Plus cleans up a bunch of elements in the SVG images that were
invisible to reduce file size.