[#3305] Add ability for enchantments to add rider items
Adds rider items referenced by UUID that are added to a parent
actor when one of their items is enchanted and removed when it is
disenchanted.

Uses a fallback for V11 which displays links to the rider items,
but doesn't allow adding or removing them except through the API.
In V12 the interface uses the new `<document-tags>` custom element
for editing.