[#3444] Re-implement summoned item modifications using enchantments
Rather than changing items on summoned actors directly, this now
uses an enchantment with a fixed ID for those changes to ensure
that it can be properly updated on re-summoning linked actors.

In order to get this to work on all actor types this removes the
restriction on applying enchantments. Enchantments can still only
be added manually to item types marked `enchantable`, but any
enchantments created programatically on those items can be properly
applied.