[#3441] Removed leading / on css url() and on icon in config.mjs
There was an issue where a leading slash (/) was causing the url() css to use a absolute url path instead of a relative path. This worked without a routePrefix because the url path was the same if for both absolute and relative, however with a routePrefix we need to preserve the prefix, this was done by removing the leading slash for icons paths and in the case of a less defined url a leading ../../ was added to remove the /systems/dnd5e/ that would be added to the relative path.
Closes #3441