[#3461] Added Overridden _onClickRight in token.mjs
There was an issue where during token placement if the user right clicked we expect the token placement to end, however since the token preview is under the cursor most of the time during placement we were activating the _onClickRight for the token which then does not activate the canvas.app.view.oncontextmenu. This was fix by overriding the Token5e _onClickRight so that if the canvas.app.view.oncontextmenu event is defined, it will call that event otherwise it will call the normal _onClickRight for Token which activates the hud.
Closes #3461