[#3623] Fix vehicle currency weight & merge encumbrance calcs
Currency on vehicles we being treated as if it were measured in
tons rather than pounds, so this adds a conversion to ensure the
currently always ends up in the correct weight units.

Also removed some duplicate encumbrance calculation from vehicle
sheets so there is only once set of encumbrance data for vehicles.
This required passing a method into `prepareEncumbrance` that
allows individual actor types to determine what items are included
in the calculation.